Add files via upload
This is a python script that takes the supplier price book and cleans it for use in the current price book.

The supplier provides a price book in excel that is produced as a report so there are significant issues when copying it into a price book for pricing projects. This script provides an output that is optimal for copying and insert into price books.